pyproject.toml
[project]
name = "acceleratedbinnedstatistic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Marco Celoria", email = "celoria.marco@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "cupy-cuda12x==13.6.0",
    "cutensor-cu12==2.3.1",
    "fastrlock==0.8.3",
    "flake8>=7.3.0",
    "iniconfig==2.3.0",
    "isort>=7.0.0",
    "mpi4py==4.1.0",
    "numpy==2.4.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "pygments==2.19.2",
    "pytest==9.0.2",
    "scipy>=1.16.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

#[build-system]
#requires = ["uv_build>=0.9.22,<0.10.0"]
#build-backend = "uv_build"

[tool.hatch.build.targets.wheel]
packages = ["src/acceleratedbinnedstatistic"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
multi_line_output = 3 

